Switch activity when the user is login + Ringo activities accessible from the home view
