+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/ui/home/HomeFragment.kt
@@ -1,5 +1,6 @@
 package com.example.circuitmessing.ui.home
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -8,6 +9,8 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
+import com.example.circuitmessing.ActivityRingo
+import com.example.circuitmessing.MainActivity
 import com.example.circuitmessing.R
 import com.example.circuitmessing.databinding.FragmentHomeBinding
 
@@ -31,6 +34,19 @@ class HomeFragment : Fragment() {
                 ViewModelProviders.of(this).get(HomeViewModel::class.java)
         _homeView = FragmentHomeBinding.inflate(inflater, container, false)
         val view = homeView.root
+
+
+
+        homeView.homeRingoButton.setOnClickListener {
+            val intent = Intent(this.context, ActivityRingo::class.java)
+            startActivity(intent)
+            this.activity?.finish()
+        }
+
+
+
+
+
         return view
 
     }
