+++ CircuitMessing/app/src/main/res/drawable/home_maker_design.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="#B3DA354B"/>
+    <solid android:color="#B3E3384D"/>
     <stroke android:color="#C6001A" android:width="2dp"/>
     <corners android:radius="20dp"/>
