+++ CircuitMessing/app/src/main/res/drawable/home_nibble_design.xml
@@ -2,8 +2,8 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="#B3E19800"/>
-    <stroke android:color="#C78600" android:width="2dp"/>
+    <solid android:color="#B3F1592A"/>
+    <stroke android:color="#F1592A" android:width="2dp"/>
     <corners android:radius="20dp"/>
 
 </shape>
