+++ CircuitMessing/app/src/main/res/drawable/home_ringo_design.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="#B33E87E8"/>
+    <solid android:color="#B31045BA"/>
     <stroke android:color="#00449F" android:width="2dp"/>
     <corners android:radius="20dp"/>
