+++ CircuitMessing/app/src/main/res/layout/fragment_login_fragment.xml
@@ -22,7 +22,8 @@
             <com.google.android.material.textfield.TextInputEditText
                 android:id="@+id/login_username_input"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent" />
+                android:layout_height="match_parent"
+                android:singleLine="true"/>
 
         </com.google.android.material.textfield.TextInputLayout>
